Add dollar fn operator
